+++ weed/placement/placement_test.go
@@ -1,4 +1,4 @@
-package shell
+package placement
 
 import (
 	"testing"
